# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация",
                "description": "Login by email, password",
                "operationId": "login",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Возращает JWT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "nfrwkJFJIOWE.FJIO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Возвращет данные user",
                "description": "Возвращет данные user по jwt",
                "operationId": "me",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возращает пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "User": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация",
                "description": "регистрация нового пользователя",
                "operationId": "register",
                "requestBody": {
                    "description": "передавать данные пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "first_name",
                                    "last_name",
                                    "middle_name",
                                    "inn",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Денис"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Бугаков"
                                    },
                                    "middle_name": {
                                        "type": "string",
                                        "example": "Юрьевич"
                                    },
                                    "inn": {
                                        "type": "string",
                                        "format": "inn",
                                        "example": "1092364000695"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь зарегистрирован"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Регистрация не удалась",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Такой пользователь уже существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User model",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "default": 1
                    },
                    "uuid": {
                        "title": "uuid",
                        "description": "uuid",
                        "type": "string",
                        "format": "placed",
                        "default": "8d587133-180c-3dfd-acc1-f5ab35e10730"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "format": "placed",
                        "default": "edeckow@example.org"
                    },
                    "first_name": {
                        "title": "first_name",
                        "description": "first_name",
                        "type": "string",
                        "format": "placed",
                        "default": "Pearline"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "last_name",
                        "type": "string",
                        "format": "placed",
                        "default": "Price"
                    },
                    "middle_name": {
                        "title": "middle_name",
                        "description": "middle_name",
                        "type": "string",
                        "format": "placed",
                        "default": "Eleonore"
                    },
                    "inn": {
                        "title": "inn",
                        "description": "inn",
                        "type": "string",
                        "format": "placed",
                        "default": "496625159283"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}